client/requirements.txt
SQLAlchemy==1.4.47
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
coverage==7.8.0
fastapi==0.95.2
greenlet==3.1.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
packaging==24.2
psycopg2==2.9.10
pydantic==1.10.7
pytest==8.3.5
pytest-cov==6.1.1
starlette==0.27.0
typing_extensions==4.13.2
uvicorn==0.22.0
python-dotenv==1.0.1
